Moving collection logic into a trait
